dmc implemented + apu fixes + switch to float samples
right, I did a lot here:

1) implemented the DMC. It was a bit tricky, but not as bad as I
had initially thought. I should implement CPU stall better, but w/e

2) I figured out why my pulses sounded like shit! It's because I
accidentally used sweep.val instead of sweep_val, i.e: assigning to
the union on every sweep clock! I fixed that, adn renamed sweep.val
to sweep._val so I don't make that mistake in the future.

3) Why even use `short` samples? It's 2018, we all have FPUs!

---

I also experimented with switching to SDL_QueueAudio instead of
the current Sounds_Queue.h thing I from Blargg, but there seems to
be a gradual desync issue when I use it...

Also, I found that bumping the sample-rate up greatly improved the
audio-quality, espeically on pulses. iirc, it's because I am doing
sampling in a shitty way, when I should be sampling more around key
inflection points.

See http://www.slack.net/~ant/bl-synth/ for what I mean...
(that's how Blargg's nes_snd_emu does it, and it sounds awesome)

+++ src/ui/SDL2/output_nes.cc
@@ -39,9 +39,10 @@ void SDL_GUI::step_nes() {
  */
 void SDL_GUI::output_nes() {
   // output audio!
-  short* samples = nullptr;
+  float* samples = nullptr;
   uint   count = 0;
   this->nes.console.getAudiobuff(samples, count);
+  // SDL_QueueAudio(this->sdl.nes_audiodev, samples, count * sizeof(float));
   if (count) this->sdl.nes_sound_queue.write(samples, count);
 
   // output video!
